refactor(content wrappers): move into their own namespace folder
