+++ src/modules/ContentWrappers/Sidebar.js
@@ -6,10 +6,10 @@ import { withRouter } from 'react-router-dom'
 import React from 'react'
 import propTypes from 'prop-types'
 
-import { FolderClosed } from './icons/FolderClosed'
+import { FolderClosed } from '../icons/FolderClosed'
 import { sectionPropType } from './Sidebar/sectionPropType'
 import { SidebarMenuItem } from './Sidebar/SidebarMenuItem'
-import styles from './Sidebar/styles.module.css'
+import styles from './Sidebar.module.css'
 
 const SidebarComponent = ({ sections, location, push }) => {
     const filteredSections = sections.filter(
