+++ src/modules/ContentWrappers/Sidebar/sectionPropType.js
