+++ src/pages/Overview.js
@@ -3,8 +3,8 @@ import React from 'react'
 import { Container } from '../modules/layout/Container'
 import { Content as ContentLayout } from '../modules/layout/Content'
 import { SideBar as SideBarLayout } from '../modules/layout/SideBar'
-import { Content } from '../modules/Content'
-import { Sidebar } from '../modules/Sidebar'
+import { MainContent } from '../modules/ContentWrappers/MainContent'
+import { Sidebar } from '../modules/ContentWrappers/Sidebar'
 import { CardMenu } from '../modules/CardMenu/CardMenu'
 import { sectionOrder } from '../constants/sectionOrder'
 
@@ -19,9 +19,9 @@ export const Overview = ({ name, match }) => {
             </SideBarLayout>
 
             <ContentLayout>
-                <Content>
+                <MainContent>
                     <CardMenu sections={sections} />
-                </Content>
+                </MainContent>
             </ContentLayout>
         </Container>
     )
