+++ src/pages/dataElement/DataElementList.js
@@ -4,9 +4,9 @@ import { Container } from '../../modules/layout/Container'
 import { Content as ContentLayout } from '../../modules/layout/Content'
 import { SideBar as SideBarLayout } from '../../modules/layout/SideBar'
 
-import { Content } from '../../modules/Content'
+import { MainContent } from '../../modules/ContentWrappers/MainContent'
+import { Sidebar } from '../../modules/ContentWrappers/Sidebar'
 import { PageHeadline } from '../../modules/PageHeadline'
-import { Sidebar } from '../../modules/Sidebar'
 import { dataElementSections } from '../../constants/sections'
 import { groups } from '../../constants/groups'
 import { sectionOrder } from '../../constants/sectionOrder'
@@ -18,12 +18,12 @@ export const DataElementList = () => (
         </SideBarLayout>
 
         <ContentLayout>
-            <Content>
+            <MainContent>
                 <PageHeadline>
                     {dataElementSections.dataElement.name}
                 </PageHeadline>
                 <span>Data elements list...</span>
-            </Content>
+            </MainContent>
         </ContentLayout>
     </Container>
 )
